requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
python-dotenv==0.9.1
Werkzeug==0.14.1
flask-cors==3.0.8
flask-sqlalchemy==2.4.0
flask-migrate==2.5.2
pymysql==0.9.3